Move after hook to events.rs
